add shell and message

+++ src/com/baoxue/task/task/DeletePackageTaskItem.java
@@ -3,6 +3,7 @@ package com.baoxue.task.task;
 import java.util.Map;
 
 import android.content.Intent;
+import android.util.Log;
 
 import com.baoxue.task.CrashApplication;
 import com.baoxue.task.common.Utility;
@@ -10,7 +11,15 @@ import com.baoxue.task.update.AppInfo;
 
 public class DeletePackageTaskItem extends TaskItem implements PackageItem {
 
+	private final static String ACTION_INSTALL = "baoxue.action.INSTALL_PACKAGES";
+	private final static String ACTION_DELETE = "baoxue.action.DELETE_PACKAGES";
+	private final static String EXTRA_SHOW_DIALOG = "show_dialog"; // boolean
+	private final static String EXTRA_PATH = "path"; // stirng
+	private final static String EXTRA_PACKAGE_NAME = "packagename"; // stirng
+
 	String packageName;
+	private long time;
+	private final String TAG = "DeletePackageTaskItem:" + this.hashCode();
 
 	public DeletePackageTaskItem(String packageName) {
 
@@ -24,16 +33,28 @@ public class DeletePackageTaskItem extends TaskItem implements PackageItem {
 	@Override
 	public void execute() {
 		setState(TaskItem.STATE_RUNNING);
+		time = System.currentTimeMillis();
 		Map<String, AppInfo> apps = PackageService.getPackage(CrashApplication
 				.getCurrent().getPackageManager());
 		AppInfo appInfo = apps.get(packageName);
-		String apkPath = appInfo.getApkPath();
-		String dataPath = appInfo.getDataDir();
-		String cmd = String.format("rm %s", apkPath);
-		Utility.runCommand(cmd);
-		cmd = String.format("rm -r %s", dataPath);
-		Utility.runCommand(cmd);
-		PackageService.reset();
+		if (appInfo != null) {
+			if (!appInfo.isSystemUpdateApp() && appInfo.isSystemApp()) {
+				String apkPath = appInfo.getApkPath();
+				String dataPath = appInfo.getDataDir();
+				String cmd = String.format("rm %s", apkPath);
+				Utility.runCommand(cmd);
+				cmd = String.format("rm -r %s", dataPath);
+				Utility.runCommand(cmd);
+			} else {
+				Intent i = new Intent(ACTION_DELETE);
+				i.putExtra(EXTRA_SHOW_DIALOG, false);
+				i.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+				i.putExtra(EXTRA_PACKAGE_NAME, packageName);
+				CrashApplication.getCurrent().sendBroadcast(i);
+			}
+		} else {
+			setState(TaskItem.STATE_COMPLATE);
+		}
 	}
 
 	@Override
@@ -43,7 +64,12 @@ public class DeletePackageTaskItem extends TaskItem implements PackageItem {
 
 	@Override
 	public void checkTimeout() {
-
+		if (getState() == TaskItem.STATE_RUNNING) {
+			if ((System.currentTimeMillis() - time) > 60000) {
+				Log.d(TAG, "timeOut");
+				setState(TaskItem.STATE_COMPLATE);
+			}
+		}
 	}
 
 	@Override
